Broken access fixed

+++ src/main/java/com/userdetails/service/UserDetailsServiceImpl.java
@@ -7,8 +7,11 @@ import com.userdetails.repository.UserDetailsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
+
+import java.io.IOException;
 
 @Service
 public class UserDetailsServiceImpl implements UserDetailsService {
@@ -22,6 +25,9 @@ public class UserDetailsServiceImpl implements UserDetailsService {
     @Autowired
     private WebClient.Builder webClientBuilder;
 
+    @Autowired
+    JwtUtil jwtUtil;
+
     public UserDetailsServiceImpl(KafkaTemplate<String, String> template) {
         this.template = template;
     }
@@ -31,6 +37,13 @@ public class UserDetailsServiceImpl implements UserDetailsService {
         return repository.getOne(id);
 
     }
+
+    private Long getIdFromJwtToken(String token){
+        return jwtUtil.getIdFromToken(token);
+    }
+
+
+
 
     private String registerUserInAuthService(UserDTO user) throws JsonProcessingException {
         String url = String.format("http://34.91.3.123:8081/register");
@@ -49,6 +62,7 @@ public class UserDetailsServiceImpl implements UserDetailsService {
     }
 
     private String deleteUserInAuthService(Long id){
+
         try {
             template.send("deleteUser", id.toString());
             return "success";
@@ -81,11 +95,15 @@ public class UserDetailsServiceImpl implements UserDetailsService {
     }
 
     @Override
-    public String deleteUserById(Long id) {
-        if(deleteUserInAuthService(id).equals("success"))
-        {
-            repository.deleteById(id);
-            return "success";
+    public String deleteUserById(Long id, String token) {
+        String jwtToken = token.substring(7);
+        Long userId = getIdFromJwtToken(jwtToken);
+        if(userId.equals(id)) {
+            if (deleteUserInAuthService(id).equals("success")) {
+                repository.deleteById(id);
+                return "success";
+            }
+            return "No permission";
         }
         return "error";
     }
